// src/types/commonInterface.ts
export type ButtonType = "primary" | "secondary" | "primary_link" | "secondary_link" | "primary_outline" | "secondary_outline";

export enum ButtonTypes {
    PRIMARY = "primary",
    SECONDARY = "secondary",
    PRIMARY_LINK = "primary_link",
    SECONDARY_LINK = "secondary_link",
    PRIMARY_OUTLINE = "primary_outline",
    SECONDARY_OUTLINE = "secondary_outline"
}

export enum EventNameTypes {
    BIRTHDAY = "Birthday",
    ANNIVERSARY = "Anniversary"
}